streamlit/requirements.txt
pandas==1.5.3
streamlit==1.19.0
torch==1.13.1
torchvision==0.14.1
pillow==9.4.0
wget==3.2
keras==2.10.0
face-detection==0.2.2